chore: revert to single file
