+++ test/test_binding.js
@@ -1,0 +1,235 @@
+const keyring = require("../src/node");
+const assert = require("assert");
+
+// OS Store tests
+assert(keyring.OsStore.set, "keyring.OsStore.set is undefined");
+assert(keyring.OsStore.get, "keyring.OsStore.get is undefined");
+assert(keyring.OsStore.list, "keyring.OsStore.list is undefined");
+assert(keyring.OsStore.remove, "keyring.OsStore.remove is undefined");
+
+const serviceName = "archwayhq";
+const keyName = "keyring2";
+const data = "hello world 2 - 1";
+const success = "success";
+
+function testSetOsStore() {
+  const result = keyring.OsStore.set(serviceName, keyName, data);
+  console.log("keyring.OsStore.set returns:", success);
+  assert.strictEqual(result, success, "Unexpected value returned");
+}
+
+function testGetOsStore() {
+  const result = keyring.OsStore.get(serviceName, keyName);
+  console.log("keyring.OsStore.get returns:", result);
+  assert.strictEqual(result, data, "Unexpected value returned");
+}
+
+function testListOsStore() {
+  const result = keyring.OsStore.list(serviceName);
+  console.log("keyring.OsStore.list returns:", result);
+  assert(
+    Array.isArray(result) && result.includes(keyName),
+    "Unexpected value returned"
+  );
+}
+
+function testRemoveOsStore() {
+  const result = keyring.OsStore.remove(serviceName, keyName);
+  console.log("keyring.OsStore.remove returns:", result);
+
+  const list = keyring.OsStore.list(serviceName);
+  assert(
+    Array.isArray(list) && !list.includes(keyName),
+    "Unexpected value returned"
+  );
+}
+
+// Encrypted File Store tests
+assert(keyring.FileStore.set, "keyring.FileStore.set is undefined");
+assert(keyring.FileStore.get, "keyring.FileStore.get is undefined");
+assert(keyring.FileStore.list, "keyring.FileStore.list is undefined");
+assert(keyring.FileStore.remove, "keyring.FileStore.remove is undefined");
+
+const fileSaveDir = "~/";
+const fileNameFile = "file-keyring2";
+const dataForFile = "hello world file 2";
+const filePassword = "password123";
+
+function testSetFileStore() {
+  const result = keyring.FileStore.set(
+    fileSaveDir,
+    fileNameFile,
+    dataForFile,
+    filePassword
+  );
+  console.log("keyring.FileStore.set returns:", result);
+  assert.strictEqual(result, success, "Unexpected value returned");
+}
+
+function testGetFileStore() {
+  const dataResult = keyring.FileStore.get(
+    fileSaveDir,
+    fileNameFile,
+    filePassword
+  );
+  console.log("keyring.FileStore.get returns:", dataResult);
+  assert.strictEqual(dataResult, dataForFile, "Unexpected value returned");
+}
+
+function testListFileStore() {
+  const result = keyring.FileStore.list(fileSaveDir);
+  console.log("keyring.FileStore.list returns:", result);
+  assert(
+    Array.isArray(result) && result.includes(fileNameFile),
+    "Unexpected value returned"
+  );
+}
+
+function testRemoveFileStore() {
+  const result = keyring.FileStore.remove(fileSaveDir, fileNameFile);
+  console.log("keyring.FileStore.remove returns:", result);
+
+  const list = keyring.FileStore.list(fileSaveDir);
+  assert(
+    Array.isArray(list) && !list.includes(fileNameFile),
+    "Unexpected value returned"
+  );
+}
+
+// Unencrypted File Store tests
+assert(
+  keyring.UnencryptedFileStore.set,
+  "keyring.UnencryptedFileStore.set is undefined"
+);
+assert(
+  keyring.UnencryptedFileStore.get,
+  "keyring.UnencryptedFileStore.get is undefined"
+);
+assert(
+  keyring.UnencryptedFileStore.list,
+  "keyring.UnencryptedFileStore.list is undefined"
+);
+assert(
+  keyring.UnencryptedFileStore.remove,
+  "keyring.UnencryptedFileStore.remove is undefined"
+);
+
+const dataForUnencryptedFile = "hello world file 3";
+
+function testSetUnencryptedFileStore() {
+  const result = keyring.UnencryptedFileStore.set(
+    fileSaveDir,
+    fileNameFile,
+    dataForUnencryptedFile
+  );
+  console.log("keyring.UnencryptedFileStore.set returns", result);
+  assert.strictEqual(result, success, "Unexpected value returned");
+}
+
+function testGetUnencryptedFileStore() {
+  const dataResult = keyring.UnencryptedFileStore.get(
+    fileSaveDir,
+    fileNameFile
+  );
+  console.log("keyring.UnencryptedFileStore.get returns", dataResult);
+  assert.strictEqual(
+    dataResult,
+    dataForUnencryptedFile,
+    "Unexpected value returned"
+  );
+}
+
+function testListUnencryptedFileStore() {
+  const result = keyring.UnencryptedFileStore.list(fileSaveDir);
+  console.log("keyring.UnencryptedFileStore.list returns:", result);
+  assert(
+    Array.isArray(result) && result.includes(fileNameFile),
+    "Unexpected value returned"
+  );
+}
+
+function testRemoveUnencryptedFileStore() {
+  const result = keyring.UnencryptedFileStore.remove(fileSaveDir, fileNameFile);
+  console.log("keyring.UnencryptedFileStore.remove returns:", result);
+
+  const list = keyring.UnencryptedFileStore.list(fileSaveDir);
+  assert(
+    Array.isArray(list) && !list.includes(fileNameFile),
+    "Unexpected value returned"
+  );
+}
+
+assert.doesNotThrow(
+  testSetOsStore,
+  undefined,
+  "testSetOsStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testGetOsStore,
+  undefined,
+  "testGetOsStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testListOsStore,
+  undefined,
+  "testListOsStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testRemoveOsStore,
+  undefined,
+  "testRemoveOsStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testSetFileStore,
+  undefined,
+  "testSetFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testGetFileStore,
+  undefined,
+  "testGetFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testListFileStore,
+  undefined,
+  "testListFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testRemoveFileStore,
+  undefined,
+  "testRemoveFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testSetUnencryptedFileStore,
+  undefined,
+  "testSetUnencryptedFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testGetUnencryptedFileStore,
+  undefined,
+  "testGetUnencryptedFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testListUnencryptedFileStore,
+  undefined,
+  "testListUnencryptedFileStore threw an expection"
+);
+console.log("");
+assert.doesNotThrow(
+  testRemoveUnencryptedFileStore,
+  undefined,
+  "testRemoveUnencryptedFileStore threw an expection"
+);
+console.log("");
+
+console.log("Tests passed- everything looks OK!");
